Add shortcut key to activate Wallabag. Resolves #42
A new preference option was added with a default of accel-alt-s. The
last character can be modified in the preferences. When modified it will
remove the previous shortcut key and if the new character is valid then
it will add the new shortcut. When the shortcut key is used Wallabag will
be activated.